+++ test/mocks/data/optionsResponse.js
@@ -1,0 +1,14 @@
+const optionsResponse = [
+  {
+    blogname: 'Test',
+    blogdescription: 'Test response without id',
+  },
+];
+
+optionsResponse._paging = {
+  total: 1,
+  totalPages: 1,
+};
+
+export default optionsResponse;
+